refactored results logic into new folder
