+++ frontend/vite/src/Components/Navbar/Navbar.jsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {
   Link
 } from 'react-router-dom';
-import BackendActor from '../../Server/BackendActor/backend-actor';
+import {BackendActor} from '../../Server/BackendActor/backend-actor';
 
 import {
   Flex,
